# backend/api/openapi3/api.yaml
openapi: "3.0.3"
info:
  title: TS Community OS API
  description: Backend API for the TS Community OS Telegram Mini App
  version: 0.3.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/auth/telegram:
    post:
      operationId: authTelegram
      summary: Authenticate via Telegram initData
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid init data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/school:
    post:
      operationId: authSchool
      summary: Verify student via school credentials
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolAuthRequest"
      responses:
        "200":
          description: School verification successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid school credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/news:
    get:
      operationId: listNews
      summary: List news articles
      tags: [news]
      parameters:
        - name: tag
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of news
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/News"
    post:
      operationId: createNews
      summary: Create a news article (admin only)
      tags: [news]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsCreateRequest"
      responses:
        "201":
          description: News created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/news/{id}:
    get:
      operationId: getNews
      summary: Get a single news article
      tags: [news]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateNews
      summary: Update a news article (admin only)
      tags: [news]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsCreateRequest"
      responses:
        "200":
          description: News updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNews
      summary: Delete a news article (admin only)
      tags: [news]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: News deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Hackathons ──────────────────────────────────────────
  /api/hackathons:
    get:
      operationId: listHackathons
      summary: List hackathons
      tags: [hackathons]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, past]
      responses:
        "200":
          description: List of hackathons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hackathon"
    post:
      operationId: createHackathon
      summary: Create a hackathon (admin only)
      tags: [hackathons]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HackathonCreateRequest"
      responses:
        "201":
          description: Hackathon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hackathon"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/hackathons/{id}:
    get:
      operationId: getHackathon
      summary: Get a single hackathon
      tags: [hackathons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Hackathon details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hackathon"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteHackathon
      summary: Delete a hackathon (admin only)
      tags: [hackathons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Hackathon deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/hackathons/{id}/apply:
    post:
      operationId: applyToHackathon
      summary: Apply to a hackathon
      tags: [hackathons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HackathonApplyRequest"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HackathonApplication"
        "409":
          description: Already applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/hackathons/{id}/applications:
    get:
      operationId: listHackathonApplications
      summary: List applications for a hackathon (admin only)
      tags: [hackathons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HackathonApplication"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Attendance ──────────────────────────────────────────
  /api/attendance/check-in:
    post:
      operationId: attendanceCheckIn
      summary: Check in a student via QR (admin only)
      tags: [attendance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInRequest"
      responses:
        "201":
          description: Check-in recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already checked in today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/clubs:
    get:
      operationId: listClubs
      summary: List all clubs
      tags: [clubs]
      responses:
        "200":
          description: List of clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Club"
    post:
      operationId: createClub
      summary: Create a club (admin only)
      tags: [clubs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClubCreateRequest"
      responses:
        "201":
          description: Club created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/clubs/{id}:
    get:
      operationId: getClub
      summary: Get a single club
      tags: [clubs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Club details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClub
      summary: Delete a club (admin only)
      tags: [clubs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Club deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/clubs/{id}/join:
    post:
      operationId: joinClub
      summary: Join a club
      tags: [clubs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
        "409":
          description: Already a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/clubs/{id}/leave:
    delete:
      operationId: leaveClub
      summary: Leave a club
      tags: [clubs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Left the club

  # ── Student Government ──────────────────────────────────
  /api/gov:
    get:
      operationId: listGovMembers
      summary: List government members
      tags: [gov]
      responses:
        "200":
          description: Government members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GovMember"
    post:
      operationId: createGovMember
      summary: Add a government member (admin only)
      tags: [gov]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovMemberCreateRequest"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovMember"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/gov/{id}:
    delete:
      operationId: deleteGovMember
      summary: Remove a government member (admin only)
      tags: [gov]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Member removed
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    AuthRequest:
      type: object
      required: [init_data]
      properties:
        init_data:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    SchoolAuthRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, telegram_id, role]
      properties:
        id:
          type: integer
          format: int64
        telegram_id:
          type: integer
          format: int64
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        photo_url:
          type: string
        role:
          type: string
          enum: [guest, student, club_leader, admin]
        school_login:
          type: string
        school_level:
          type: integer
        school_xp:
          type: integer
          format: int64
        audit_ratio:
          type: number
          format: float
        coins:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    News:
      type: object
      required: [id, title, content, tag, created_at]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        image_url:
          type: string
        tag:
          type: string
        author_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NewsCreateRequest:
      type: object
      required: [title, content, tag]
      properties:
        title:
          type: string
        content:
          type: string
        image_url:
          type: string
        tag:
          type: string
...